Adds a hemispheric JH integral to WAM.
By-timestep output available in RUNDIR with nam_phy namelist's
output_jh_integral. Still need to decide if those integrals should
be made available to physics at the next timestep (and what to
do at the first timestep), and if this functionality should be
controlled by a namelist switch, all of which will require additional
attention.