Same timestep JH integral.
I don't know if this code will work but I sure hope it does.

A flat JH factor of 2 is used in idea_ion, and physics continues as
normal. dtdt_jh (post-IPE merge dtdt from JH subroutines) is passed
out of idea_phys into gloopb and then do_physics_one_step, where the
JH integral is immediately calculated per normal. Then we can use this
value in conjunction with the dtdt_jh array to come up with an empirical
jh_fac, and subtract 2 to account for the flat factor used in idea_ion.
We should get an accurate temp field as a result without affecting
physics.